backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.3.0

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & Caching
redis==5.0.1
hiredis==2.2.3

# Task Queue
celery==5.3.4
flower==2.0.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
scipy==1.11.4

# AI/ML Libraries
basic-pitch==0.2.5
demucs==4.0.1
essentia==2.1b6.dev1110
music21==9.1.0
mingus==0.6.1

# YouTube & External APIs
yt-dlp==2023.11.16
spotipy==2.23.0
musicbrainzngs==0.7.1
google-api-python-client==2.108.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
